rtlamr2mqtt-addon/requirements.txt
paho-mqtt==2.1.0
pyyaml==6.0.2
requests==2.32.3
pyusb==1.3.1
